+++ Genius/Models/Settings/Defaults.Keys.swift
@@ -10,6 +10,7 @@ extension Defaults.Keys {
 	static let useTextInsteadOfSymbols = Key<Bool>("UseTextInsteadOfSymbols", default: { Defaults[.interfaceMode] <= .simple })
 	static let interfaceMode = Key<Settings.InterfaceMode>("InterfaceMode", default: .normal)
 	static let disableLiquidGlass = Key<Bool>("DisableLiquidGlass", default: false)
+	static let hideIconsInMenuBar = Key<Bool>("HideIconsInMenuBar", default: { Defaults[.disableLiquidGlass] })
 
 	static let betaUpdates = Key<Bool>("BetaUpdates", default: false)